ads code refactoring
