+++ RemixedDungeon/src/googlePlay/java/com/nyrds/pixeldungeon/support/AdMobInterstitialProvider.java
@@ -1,13 +1,7 @@
 package com.nyrds.pixeldungeon.support;
 
-import android.annotation.SuppressLint;
-import android.graphics.Color;
-
 import androidx.annotation.NonNull;
 
-import com.google.android.gms.ads.AdListener;
-import com.google.android.gms.ads.AdSize;
-import com.google.android.gms.ads.AdView;
 import com.google.android.gms.ads.FullScreenContentCallback;
 import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.interstitial.InterstitialAd;
@@ -16,11 +10,10 @@ import com.nyrds.pixeldungeon.ml.R;
 import com.watabou.noosa.Game;
 import com.watabou.noosa.InterstitialPoint;
 
-public class AdMobComboProvider implements AdsUtilsCommon.IInterstitialProvider, AdsUtilsCommon.IBannerProvider {
+public class AdMobInterstitialProvider implements AdsUtilsCommon.IInterstitialProvider {
     private static InterstitialAd mInterstitialAd = null;
-    private AdView adView;
 
-    AdMobComboProvider() {
+    AdMobInterstitialProvider() {
         Game.instance().runOnUiThread(this::requestNewInterstitial);
     }
 
@@ -51,7 +44,7 @@ public class AdMobComboProvider implements AdsUtilsCommon.IInterstitialProvider,
     public void showInterstitial(final InterstitialPoint ret) {
         Game.instance().runOnUiThread(() -> {
             if (mInterstitialAd == null) {
-                AdsUtilsCommon.interstitialFailed(AdMobComboProvider.this, ret);
+                AdsUtilsCommon.interstitialFailed(AdMobInterstitialProvider.this, ret);
                 return;
             }
 
@@ -70,32 +63,9 @@ public class AdMobComboProvider implements AdsUtilsCommon.IInterstitialProvider,
 
     }
 
-    @SuppressLint("MissingPermission")
-    @Override
-    public void displayBanner() {
-        adView = new AdView(Game.instance());
-        adView.setAdUnitId(Game.getVar(R.string.easyModeAdUnitId));
-        adView.setBackgroundColor(Color.TRANSPARENT);
-        adView.setAdListener(new AdmobBannerListener());
-        adView.setAdSize(AdSize.SMART_BANNER);
-        adView.loadAd(AdMob.makeAdRequest());
-    }
-
     @Override
     public boolean isReady() {
-        return true;
+        return mInterstitialAd!=null;
     }
 
-    private class AdmobBannerListener extends AdListener {
-
-        @Override
-        public void onAdLoaded() {
-            Ads.updateBanner(adView);
-        }
-
-        @Override
-        public void onAdFailedToLoad(LoadAdError reason) {
-            AdsUtilsCommon.bannerFailed(AdMobComboProvider.this);
-        }
-    }
 }
